==> Amanda/2_analise_exploratoria.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from collections import Counter

import plotly.express as px
from plotly.subplots import make_subplots
import plotly.graph_objects as go
from math import ceil

df = pd.read_csv('Dados_Com.csv', index_col=0)

'''# Contar a frequência de cada editora
contagem_publisher = Counter(PublisherTotal['boardgamepublisher'])

# Criar um DataFrame com a contagem
df_contagem_publisher = pd.DataFrame({'Publisher': list(contagem_publisher.keys()), 'Contagem_publisher': list(contagem_publisher.values())})
df_contagem_publisher.sort_values(by="Contagem_publisher", ascending=False, inplace=True)

# Criar o treemap
fig = px.treemap(df_contagem_publisher.head(20), path=['Publisher'], values='Contagem_publisher',
                 title='20 maiores editoras')
fig.show()'''

df['total_time'] = df['order_metric_collected_time'] + df['order_metric_paused_time'] + df['order_metric_production_time'] + df['order_metric_walking_time'] + df['order_metric_expediton_speed_time'] + df['order_metric_transit_time'] + df['order_metric_cycle_time']

# Agrupando o DataFrame pelas cidades e seus estados correspondentes
grouped_data = df.groupby(['hub_city', 'hub_state']).size().reset_index(name='Count')

# Criando o treemap
fig = px.treemap(grouped_data, path=['hub_state', 'hub_city'], values='Count',
                 title='Cidades em que Hack Eats está presente')
fig.show()

# Definindo os limites dos intervalos
limites = [0, 500, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, float('inf')]

# Definindo os rótulos para cada intervalo
rotulos = ['0-500', '500-1000', '1000-1500', '1500-2000', '2000-2500', '2500-3000', '3000-3500', '3500-4000', '4000-4500', '4500-5000', '>5000']

# Criando a nova coluna 'distance_range' com base nos intervalos definidos
df['distance_range'] = pd.cut(df['delivery_distance_meters'], bins=limites, labels=rotulos, right=False)

# Calculando a contagem de linhas para cada 'distance_range'
contagem_por_distancia = df.groupby(['distance_range', 'delivery_status']).size().reset_index(name='count')

# Criando o gráfico de barras com a contagem de linhas por 'distance_range' e 'delivery_status'
fig = px.bar(contagem_por_distancia, x='distance_range', y='count', color='delivery_status', 
             title='Status em função da distância')
fig.show()

A distância não afeta o cancelamento

# Calculando a contagem de linhas para cada 'channel_type'
contagem_por_canal = df.groupby(['channel_type', 'driver_type']).size().reset_index(name='count')

fig = px.histogram(contagem_por_canal, x='channel_type', y='count', color='driver_type', barmode='group')
fig.show()

# Calculando a contagem de linhas para cada 'channel_type' e 'driver_type'
contagem_por_canal = df.groupby(['channel_type', 'driver_type']).size().reset_index(name='count')

# Normalizando os valores para porcentagem
contagem_por_canal['percentage'] = contagem_por_canal.groupby('channel_type')['count'].transform(lambda x: x / x.sum() * 100)

# Criando o gráfico de histograma com porcentagens
fig = px.histogram(contagem_por_canal, x='channel_type', y='percentage', color='driver_type', 
                   barmode='group', histfunc='sum', title='Porcentagem por canal e tipo de motorista')
fig.update_yaxes(title='Porcentagem')

fig.show()

Não importa o tipo de canal, a proporção do tipo de motorista é a mesma

fig = px.scatter(df, x='delivery_distance_meters', y='total_time', title = 'Distancia por tempo de entrega')
fig.show()

O esperado era que o tempo total aumentasse linearmente com a distancia, mas nao foi o que aconteceu

# Calculando a contagem de linhas para cada 'driver_modal'
contagem_por_modal = df.groupby(['driver_modal', 'hub_city']).size().reset_index(name='count')

fig = px.histogram(contagem_por_modal, x='hub_city', y='count', color='driver_modal', barmode='group')
fig.show()

# Agrupando o DataFrame pelos modos de pagamentos
grouped_payment_method = df.groupby(['payment_method', 'delivery_status']).size().reset_index(name='Count')

# Criando o treemap
fig = px.treemap(grouped_payment_method, path=['payment_method'], values='Count',
                 title='Tipos de pagamentos')
fig.show()

# Normalizando os valores para porcentagem
grouped_payment_method['percentage'] = grouped_payment_method.groupby('payment_method')['Count'].transform(lambda x: x / x.sum() * 100)

# Criando o gráfico de histograma com porcentagens
fig = px.histogram(grouped_payment_method, x='payment_method', y='percentage', color='delivery_status', barmode='group', histfunc='sum', title='Porcentagem por metodo de pagamento e status de entrega')
fig.update_yaxes(title='Porcentagem')
fig.show()

Todas as entregas que foram pagas forma entregues

fig = px.histogram(df, x='store_name', y='order_metric_cycle_time', title='Distribuição de order_metric_cycle_time por store_name', histfunc='avg')
fig.show()

fig = px.histogram(df, x='driver_type', y='order_metric_cycle_time', title='Distribuição de order_metric_cycle_time por driver_type', histfunc='avg')

fig.show()

fig = px.histogram(df, x='payment_fee', y='order_metric_cycle_time', title='Distribuição de order_metric_cycle_time por payment_fee', histfunc='avg')

fig.show()

fig = px.histogram(df, x='payment_method', y='order_metric_cycle_time', title='Distribuição de order_metric_cycle_time por payment_method', histfunc='avg')

fig.show()

fig = px.histogram(df, x='channel_name', y='order_metric_cycle_time', title='Distribuição de order_metric_cycle_time por channel_name', histfunc='avg')

fig.show()

fig = px.histogram(df, x='order_delivery_fee', y='order_metric_cycle_time', title='Distribuição de order_metric_cycle_time por order_delivery_fee', histfunc='avg')

fig.show()
