==> Amanda/3_modelagem.ipynb <==
# Modelagem #


import pandas as pd
import numpy as np

from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
#from sklearn.linear_model import LogisticRegression
#from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score, f1_score
from sklearn.neural_network import MLPRegressor


df1 = pd.read_csv('base_limpa1.csv', index_col=0)
df2 = pd.read_csv('base_limpa2.csv', index_col=0)

df = pd.concat([df1, df2])
df = df.dropna()

df['order_date'] = pd.to_datetime(df['order_date'])
#df_date = df.copy()

# Criando colunas para facilitar análises
df['minute'] = df['order_date'].dt.minute
df['hour'] = df['order_date'].dt.hour
df['day'] = df['order_date'].dt.day
df['month'] = df['order_date'].dt.month
df['year'] = df['order_date'].dt.year
df['weekday'] = df['order_date'].dt.weekday # Domingo é 0

# Definir as variáveis independentes (características)
X = df[['hub_name', 'hub_state', 'store_name', 'hour', 'weekday', 'delivery_distance_meters', 'payment_method', 'channel_name', 'channel_type', 'order_delivery_fee']] 
X_encoded = pd.get_dummies(X, columns=['hub_name', 'hub_state', 'store_name', 'payment_method', 'channel_name', 'channel_type'])
X_not_encoded =df[['hour', 'weekday', 'delivery_distance_meters','order_delivery_fee']]

# Definir a variável dependente (o que você quer prever)
y = df['order_metric_cycle_time']

# Dividir os dados em conjuntos de treino e teste para variáveis codificadas
X_train_encoded, X_test_encoded, y_train, y_test = train_test_split(X_encoded, y, test_size=0.2, random_state=42)

# Supondo que X_not_encoded são as variáveis que não passaram pelo Dummy Encoding
X_train_not_encoded, X_test_not_encoded, _, _ = train_test_split(X_not_encoded, y, test_size=0.2, random_state=42)

# Concatenar as colunas das variáveis codificadas e não codificadas
X_train = pd.concat([X_train_encoded, X_train_not_encoded], axis=1)
X_test = pd.concat([X_test_encoded, X_test_not_encoded], axis=1)

# Instanciar o modelo de regressão linear
model = LinearRegression()

# Treinar o modelo
model.fit(X_train, y_train)

# Fazer previsões no conjunto de teste
predictions = model.predict(X_test)

# Avaliar o desempenho do modelo
mse = mean_squared_error(y_test, predictions)
rmse = np.sqrt(mse)
mae = mean_absolute_error(y_test,predictions)
r2 = r2_score(y_test, predictions)
print(f'MSE:{mse} | RMSE:{rmse} | MAE:{mae} | R2:{r2}')

#Criar e treinar o modelo de Neural Regression
model = MLPRegressor(hidden_layer_sizes=(200, 100), activation='relu', solver='adam', random_state=42)
model.fit(X_train, y_train)

# Fazer previsões no conjunto de teste
predictions = model.predict(X_test)

# Avaliar o desempenho do modelo
mse = mean_squared_error(y_test, predictions)
rmse = np.sqrt(mse)
mae = mean_absolute_error(y_test,predictions)
r2 = r2_score(y_test, predictions)
print(f'MSE:{mse} | RMSE:{rmse} | MAE:{mae} | R2:{r2}')
